🐛 Fix: 선수 세부 페이지 클라이언트-서버 컴포넌트 분리 오류 수정
